move root and hydromt_step to model/
