+++ hydromt/__init__.py
@@ -13,8 +13,6 @@ import netCDF4  # noqa: F401
 # required for accessor style documentation
 from xarray import DataArray, Dataset  # noqa: F401
 
-from hydromt.hydromt_step import hydromt_step
-
 from . import cli, gis, stats, workflows
 from .data_catalog import *
