+++ hydromt/model/__init__.py
@@ -1,8 +1,9 @@
 """HydroMT Model API."""
 
+from .hydromt_step import hydromt_step
 from .model import Model
 
-__all__ = ["Model"]
+__all__ = ["Model", "hydromt_step"]
 
 # define hydromt model entry points
 # see also hydromt.model group in pyproject.toml
