+++ hydromt/model/components/mesh.py
@@ -12,9 +12,9 @@ import xugrid as xu
 from pyproj import CRS
 from shapely.geometry import box
 
-from hydromt import hydromt_step
 from hydromt.gis.raster import GEO_MAP_COORD
 from hydromt.io.readers import read_nc
+from hydromt.model import hydromt_step
 from hydromt.model.components.base import ModelComponent
 from hydromt.model.components.spatial import SpatialModelComponent
 from hydromt.workflows.mesh import (
