+++ hydromt/model/components/spatialdatasets.py
@@ -18,11 +18,11 @@ from pandas import DataFrame
 from xarray import DataArray, Dataset
 
 from hydromt._typing.type_def import DeferedFileClose, XArrayDict
-from hydromt.hydromt_step import hydromt_step
 from hydromt.io.readers import read_nc
 from hydromt.io.writers import write_nc
 from hydromt.model.components.base import ModelComponent
 from hydromt.model.components.spatial import SpatialModelComponent
+from hydromt.model.hydromt_step import hydromt_step
 
 if TYPE_CHECKING:
     from hydromt.model.model import Model
