+++ hydromt/model/components/vector.py
@@ -12,10 +12,10 @@ from geopandas.testing import assert_geodataframe_equal
 from pyproj import CRS
 from shapely.geometry import box
 
-from hydromt import hydromt_step
 from hydromt.gis.vector import GeoDataset
 from hydromt.io.readers import read_nc
 from hydromt.io.writers import write_nc
+from hydromt.model import hydromt_step
 from hydromt.model.components.base import ModelComponent
 from hydromt.model.components.spatial import SpatialModelComponent
