+++ hydromt/model/model.py
@@ -20,19 +20,19 @@ from typing import (
 import geopandas as gdp
 from pyproj import CRS
 
-from hydromt import hydromt_step
 from hydromt._typing import StrPath
 from hydromt._typing.type_def import DeferedFileClose
 from hydromt._utils.rgetattr import rgetattr
 from hydromt._utils.steps_validator import validate_steps
 from hydromt.data_catalog import DataCatalog
+from hydromt.model import hydromt_step
 from hydromt.model.components import (
     DatasetsComponent,
     ModelComponent,
     SpatialModelComponent,
 )
+from hydromt.model.root import ModelRoot
 from hydromt.plugins import PLUGINS
-from hydromt.root import ModelRoot
 
 __all__ = ["Model"]
