+++ tests/_utils/test_steps_validator.py
@@ -1,7 +1,7 @@
 import pytest
 
-from hydromt import hydromt_step
 from hydromt._utils.steps_validator import validate_steps
+from hydromt.model import hydromt_step
 from hydromt.model.components.base import ModelComponent
 from hydromt.model.model import Model
